+++ pupil_recording_interface/__init__.py
@@ -40,6 +40,8 @@ __all__ = [
     "load_dataset",
     "load_info",
     "load_user_info",
+    "load_pldata",
+    "save_pldata",
     "write_netcdf",
     "get_gaze_mappers",
     # Readers
@@ -112,10 +114,10 @@ def load_dataset(
     Returns
     -------
     xarray.Dataset or tuple thereof
-        The recording data as a dataset or tuple thereof if both `gaze` and
-        `odometry` are specified.
+        The recording data as a dataset or tuple thereof if multiple sources
+        (`gaze`, `odometry`, ...) are specified.
     """
-    folder = Path(folder)
+    folder = Path(folder).expanduser()
     return_vals = tuple()
 
     if gaze:
@@ -149,7 +151,7 @@ def get_gaze_mappers(folder):
     set
         The set of available mappers.
     """
-    folder = Path(folder)
+    folder = Path(folder).expanduser()
     if not folder.exists():
         raise FileNotFoundError(f"No such folder: {folder}")
 
@@ -191,13 +193,13 @@ def write_netcdf(
         The source of the gyro data. Can be 'recording'.
     """
     if output_folder is None:
-        output_folder = Path(folder) / "exports"
+        output_folder = Path(folder).expanduser() / "exports"
         counter = 0
         while (output_folder / f"{counter:03d}").exists():
             counter += 1
         output_folder = output_folder / f"{counter:03d}"
     else:
-        output_folder = Path(output_folder)
+        output_folder = Path(output_folder).expanduser()
 
     if gaze is not None:
         GazeReader(folder, source=gaze).write_netcdf(
@@ -218,6 +220,42 @@ def write_netcdf(
         MotionReader(folder, "gyro", source=gyro).write_netcdf(
             filename=output_folder / "gyro.nc"
         )
+
+
+def load_pldata(folder, topic):
+    """ Load data from a .pldata file as a list of dicts.
+
+    Parameters
+    ----------
+    folder : str or pathlib.Path
+        Path to the recording folder.
+
+    topic : str
+        The topic to load, e.g. "gaze".
+
+    Returns
+    -------
+    data : list of dict
+        The loaded data.
+    """
+    return BaseReader(folder).load_pldata(topic)
+
+
+def save_pldata(folder, topic, data):
+    """ Save data from a list of dicts as a .pldata file.
+
+    Parameters
+    ----------
+    folder : str or pathlib.Path
+        Path to the recording folder.
+
+    topic : str
+        The topic to load, e.g. "gaze".
+
+    data : list of dict
+        The data to be saved.
+    """
+    BaseReader(folder).save_pldata(topic, data)
 
 
 def load_info(folder):