+++ pupil_recording_interface/process/pupil_detector.py
@@ -21,6 +21,7 @@ class PupilDetector(BaseProcess):
         camera_id=None,
         folder=None,
         record=False,
+        display=True,
         **kwargs,
     ):
         """ Constructor. """
@@ -28,6 +29,7 @@ class PupilDetector(BaseProcess):
         self.camera_id = camera_id
         self.folder = folder
         self.record = record
+        self.display = display
 
         super().__init__(**kwargs)
 
@@ -93,6 +95,30 @@ class PupilDetector(BaseProcess):
         """ Write pupil datum to disk. """
         self.writer.append(packet["pupil"])
 
+    def display_hook(self, packet):
+        """ Add pupil overlay onto frame. """
+        pupil = packet["pupil"]
+        if pupil is None:
+            # Return the attribute to avoid unnecessary waiting
+            return packet.display_frame
+
+        frame = packet["display_frame"]
+        if frame.ndim == 2:
+            frame = cv2.cvtColor(frame, cv2.COLOR_GRAY2BGR)
+
+        ellipse = pupil["ellipse"]
+        cv2.ellipse(
+            frame,
+            tuple(int(v) for v in ellipse["center"]),
+            tuple(int(v / 2) for v in ellipse["axes"]),
+            ellipse["angle"],
+            0,
+            360,  # start/end angle for drawing
+            (0, 0, 255),  # color (BGR): red
+        )
+
+        return frame
+
     def _process_packet(self, packet, block=None):
         """ Process a new packet. """
         packet.pupil = self.call(self.detect_pupil, packet, block=block)
@@ -101,6 +127,9 @@ class PupilDetector(BaseProcess):
             self.call(self.record_data, packet, block=block)
 
         packet.broadcasts.append("pupil")
+
+        if self.display:
+            packet.display_hooks.append(self.display_hook)
 
         return packet